Move the processors into their own files
This may end up being somewhat annoying as the API may not be finalized.
But hey, you have to go for it sometimes, right?
