+++ src/processor/md_processor.rs
@@ -1,0 +1,21 @@
+use std::fs::File;
+use std::io::{BufReader, Read, Write};
+use std::path::Path;
+
+use crate::processor::{Processor, ProcessorError, ProcessorKind};
+
+pub(crate) struct MdProcessor {
+    contents: BufReader<File>,
+}
+
+impl Processor for MdProcessor {
+    fn new(file: File) -> Self { MdProcessor { contents: BufReader::new(file) } }
+    fn kind(&self) -> ProcessorKind { ProcessorKind::Md }
+    fn write_to(&mut self, path: &Path) -> Result<(), ProcessorError> {
+        let mut file = File::create(path)?;
+        let mut output = String::from("<!-- Markdown HTML generated by Halogen -->\n");
+        self.contents.read_to_string(&mut output)?;
+        file.write_all(output.as_ref())?;
+        Ok(())
+    }
+}
