+++ src/processor/mod.rs
@@ -2,9 +2,19 @@ use std::convert::Infallible;
 use std::ffi::OsStr;
 use std::fmt::{Display, Formatter};
 use std::fs::File;
-use std::io::{BufReader, Error, Read, Write};
+use std::io::Error;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
+
+use crate::processor::css_processor::CssProcessor;
+use crate::processor::md_processor::MdProcessor;
+use crate::processor::none_processor::NoneProcessor;
+use crate::processor::org_processor::OrgProcessor;
+
+mod css_processor;
+mod md_processor;
+mod none_processor;
+mod org_processor;
 
 /// Represents a type that can process a given file for the web.
 trait Processor {
@@ -105,66 +115,6 @@ impl ProcessorFactory {
             ProcessorKind::Md => { Box::new(MdProcessor::new(file)) }
             ProcessorKind::Org => { Box::new(OrgProcessor::new(file)) }
         }
-    }
-}
-
-struct NoneProcessor {
-    contents: BufReader<File>,
-}
-impl Processor for NoneProcessor {
-    fn new(file: File) -> Self { NoneProcessor { contents: BufReader::new(file) } }
-    fn kind(&self) -> ProcessorKind { ProcessorKind::None }
-    fn write_to(&mut self, path: &Path) -> Result<(), ProcessorError> {
-        let mut file = File::create(path)?;
-        let mut output = String::new();
-        self.contents.read_to_string(&mut output)?;
-        file.write_all(output.as_ref())?;
-        Ok(())
-    }
-}
-
-struct CssProcessor {
-    contents: BufReader<File>,
-}
-impl Processor for CssProcessor {
-    fn new(file: File) -> Self { CssProcessor { contents: BufReader::new(file) } }
-    fn kind(&self) -> ProcessorKind { ProcessorKind::Css }
-    fn write_to(&mut self, path: &Path) -> Result<(), ProcessorError> {
-        let mut file = File::create(path)?;
-        let mut output = String::from("/* CSS generated by Halogen */\n");
-        self.contents.read_to_string(&mut output)?;
-        file.write_all(output.as_ref())?;
-        Ok(())
-    }
-}
-
-struct MdProcessor {
-    contents: BufReader<File>,
-}
-impl Processor for MdProcessor {
-    fn new(file: File) -> Self { MdProcessor { contents: BufReader::new(file) } }
-    fn kind(&self) -> ProcessorKind { ProcessorKind::Md }
-    fn write_to(&mut self, path: &Path) -> Result<(), ProcessorError> {
-        let mut file = File::create(path)?;
-        let mut output = String::from("<!-- Markdown HTML generated by Halogen -->\n");
-        self.contents.read_to_string(&mut output)?;
-        file.write_all(output.as_ref())?;
-        Ok(())
-    }
-}
-
-struct OrgProcessor {
-    contents: BufReader<File>,
-}
-impl Processor for OrgProcessor {
-    fn new(file: File) -> Self { OrgProcessor { contents: BufReader::new(file) } }
-    fn kind(&self) -> ProcessorKind { ProcessorKind::Org }
-    fn write_to(&mut self, path: &Path) -> Result<(), ProcessorError> {
-        let mut file = File::create(path)?;
-        let mut output = String::from("<!-- Org HTML generated by Halogen -->\n");
-        self.contents.read_to_string(&mut output)?;
-        file.write_all(output.as_ref())?;
-        Ok(())
     }
 }
